Use shared functionality for label stuff
