+++ libs/designsystem/checkbox/src/checkbox.component.ts
@@ -16,6 +16,7 @@ import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 
 import { IonicElementPartHelper } from '@kirbydesign/designsystem/helpers';
 import { IonCheckbox } from '@ionic/angular/standalone';
+import { inheritAriaLabelText, setAccessibleLabel } from '@kirbydesign/designsystem/shared';
 
 @Component({
   standalone: true,
@@ -88,14 +89,11 @@ export class CheckboxComponent implements AfterViewInit, ControlValueAccessor, O
       this._labelPlacement = 'start';
     }
 
-    this.inheritAriaLabelText();
+    this._labelText = inheritAriaLabelText(this.element.nativeElement);
 
-    if (this.text || this._labelText || this._hasSlottedContent) return;
-
-    const label = this.findItemLabel(this.element.nativeElement);
-    if (label) {
-      this._labelText = label.textContent;
-      label.setAttribute('aria-hidden', 'true');
+    if (!this.text && !this._labelText && !this._hasSlottedContent) {
+      // if no label has been set try to find a label in an item and use its text content
+      this._labelText = setAccessibleLabel(this.element.nativeElement);
     }
   }
 
@@ -121,25 +119,6 @@ export class CheckboxComponent implements AfterViewInit, ControlValueAccessor, O
 
   onBlur() {
     this._onTouched();
-  }
-
-  private findItemLabel(element: HTMLElement): HTMLIonLabelElement {
-    const itemEl = element.closest('kirby-item');
-    if (itemEl) {
-      return itemEl.querySelector('kirby-label');
-    }
-
-    return null;
-  }
-
-  private inheritAriaLabelText() {
-    const el = this.element.nativeElement;
-    const attribute = 'aria-label';
-    if (el.hasAttribute(attribute)) {
-      const value = el.getAttribute(attribute);
-      el.removeAttribute(attribute);
-      this._labelText = value;
-    }
   }
 
   // Initialize default ControlValueAccessor callback functions (noop)
