+++ libs/designsystem/radio/src/radio.component.ts
@@ -10,6 +10,7 @@ import {
 } from '@angular/core';
 import { IonRadio } from '@ionic/angular/standalone';
 import { IonicElementPartHelper } from '@kirbydesign/designsystem/helpers';
+import { inheritAriaLabelText, setAccessibleLabel } from '@kirbydesign/designsystem/shared';
 
 @Component({
   selector: 'kirby-radio',
@@ -65,14 +66,11 @@ export class RadioComponent implements AfterViewInit, OnInit {
       this._labelPlacement = 'start';
     }
 
-    this.inheritAriaLabelText();
+    this._labelText = inheritAriaLabelText(this.element.nativeElement);
 
-    if (this.text || this._labelText || this._hasSlottedContent) return;
-
-    const label = this.findItemLabel(this.element.nativeElement);
-    if (label) {
-      this._labelText = label.textContent;
-      label.setAttribute('aria-hidden', 'true');
+    if (!this.text && !this._labelText && !this._hasSlottedContent) {
+      // if no label has been set try to find a label in an item and use its text content
+      this._labelText = setAccessibleLabel(this.element.nativeElement);
     }
   }
 
@@ -88,24 +86,5 @@ export class RadioComponent implements AfterViewInit, OnInit {
 
   focus() {
     this.ionRadioElement && this.ionRadioElement.nativeElement.focus();
-  }
-
-  private inheritAriaLabelText() {
-    const el = this.element.nativeElement;
-    const attribute = 'aria-label';
-    if (el.hasAttribute(attribute)) {
-      const value = el.getAttribute(attribute);
-      el.removeAttribute(attribute);
-      this._labelText = value;
-    }
-  }
-
-  private findItemLabel(element: HTMLElement): HTMLIonLabelElement {
-    const itemEl = element.closest('kirby-item');
-    if (itemEl) {
-      return itemEl.querySelector('kirby-label');
-    }
-
-    return null;
   }
 }
