+++ libs/designsystem/shared/src/public_api.ts
@@ -7,3 +7,5 @@ export * from './component-loader.directive';
 export * from './dynamic-component';
 
 export * from './fit-heading/index';
+
+export * from './controls/label-helpers';
